+++ .Lib9c.Tests/Action/DailyReward0Test.cs
@@ -1,24 +1,22 @@
-namespace Lib9c.Tests.Action
+﻿namespace Lib9c.Tests.Action
 {
-    using System.Linq;
     using Libplanet;
     using Libplanet.Action;
     using Libplanet.Crypto;
     using Nekoyume;
     using Nekoyume.Action;
-    using Nekoyume.Model.Mail;
     using Nekoyume.Model.State;
     using Serilog;
     using Xunit;
     using Xunit.Abstractions;
 
-    public class DailyReward3Test
+    public class DailyReward0Test
     {
         private readonly IAccountStateDelta _initialState;
         private readonly Address _agentAddress;
         private readonly Address _avatarAddress;
 
-        public DailyReward3Test(ITestOutputHelper outputHelper)
+        public DailyReward0Test(ITestOutputHelper outputHelper)
         {
             Log.Logger = new LoggerConfiguration()
                 .MinimumLevel.Verbose()
@@ -60,7 +58,7 @@ namespace Lib9c.Tests.Action
         [Fact]
         public void Execute()
         {
-            var dailyRewardAction = new DailyReward3
+            var dailyRewardAction = new DailyReward0
             {
                 avatarAddress = _avatarAddress,
             };
@@ -68,7 +66,6 @@ namespace Lib9c.Tests.Action
             {
                 BlockIndex = 0,
                 PreviousStates = _initialState,
-                Random = new TestRandom(),
                 Rehearsal = false,
                 Signer = _agentAddress,
             });
@@ -76,22 +73,12 @@ namespace Lib9c.Tests.Action
             var gameConfigState = nextState.GetGameConfigState();
             var nextAvatarState = nextState.GetAvatarState(_avatarAddress);
             Assert.Equal(gameConfigState.ActionPointMax, nextAvatarState.actionPoint);
-            Assert.Single(nextAvatarState.mailBox);
-            var mail = nextAvatarState.mailBox.First();
-            var rewardMail = mail as DailyRewardMail;
-            Assert.NotNull(rewardMail);
-            var rewardResult = rewardMail.attachment as DailyReward.DailyRewardResult;
-            Assert.NotNull(rewardResult);
-            Assert.Single(rewardResult.materials);
-            var material = rewardResult.materials.First();
-            Assert.Equal(400000, material.Key.Id);
-            Assert.Equal(10, material.Value);
         }
 
         [Fact]
         public void ExecuteThrowFailedLoadStateException()
         {
-            var action = new DailyReward3
+            var action = new DailyReward0
             {
                 avatarAddress = _avatarAddress,
             };
