--- Lib9c/Action/DailyReward3.cs
@@ -1,115 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Linq;
-using Bencodex.Types;
-using Libplanet;
-using Libplanet.Action;
-using Nekoyume.Model.Item;
-using Nekoyume.Model.Mail;
-using Nekoyume.Model.State;
-using Nekoyume.TableData;
-
-namespace Nekoyume.Action
-{
-    [Serializable]
-    [ActionType("daily_reward2")]
-    public class DailyReward3 : GameAction
-    {
-        public Address avatarAddress;
-        public DailyRewardResult dailyRewardResult;
-        private const int rewardItemId = 400000;
-        private const int rewardItemCount = 10;
-
-        public override IAccountStateDelta Execute(IActionContext context)
-        {
-            IActionContext ctx = context;
-            var states = ctx.PreviousStates;
-            if (ctx.Rehearsal)
-            {
-                return states.SetState(avatarAddress, MarkChanged);
-            }
-            
-            var addressesHex = GetSignerAndOtherAddressesHex(context, avatarAddress);
-
-            if (!states.TryGetAgentAvatarStates(ctx.Signer, avatarAddress, out _, out AvatarState avatarState))
-            {
-                throw new FailedLoadStateException($"{addressesHex}Aborted as the avatar state of the signer was failed to load.");
-            }
-
-            var gameConfigState = states.GetGameConfigState();
-            if (gameConfigState is null)
-            {
-                throw new FailedLoadStateException($"{addressesHex}Aborted as the game config was failed to load.");
-            }
-
-            if (ctx.BlockIndex - avatarState.dailyRewardReceivedIndex >= gameConfigState.DailyRewardInterval)
-            {
-                avatarState.dailyRewardReceivedIndex = ctx.BlockIndex;
-                avatarState.actionPoint = gameConfigState.ActionPointMax;
-            }
-
-            // create item
-            var materialSheet = states.GetSheet<MaterialItemSheet>();
-            var materials = new Dictionary<Material, int>();
-            var material = ItemFactory.CreateMaterial(materialSheet, rewardItemId);
-            materials[material] = rewardItemCount;
-
-            var result = new DailyRewardResult
-            {
-                materials = materials,
-            };
-            
-            // create mail
-            var mail = new DailyRewardMail(result, 
-                                           ctx.BlockIndex, 
-                                           ctx.Random.GenerateRandomGuid(),
-                                           ctx.BlockIndex);
-
-            result.id = mail.id;
-            dailyRewardResult = result;
-            avatarState.UpdateV3(mail);
-            avatarState.UpdateFromAddItem(material, rewardItemCount, false);
-            return states.SetState(avatarAddress, avatarState.Serialize());
-        }
-
-        protected override IImmutableDictionary<string, IValue> PlainValueInternal => new Dictionary<string, IValue>
-        {
-            ["avatarAddress"] = avatarAddress.Serialize(),
-        }.ToImmutableDictionary();
-
-        protected override void LoadPlainValueInternal(IImmutableDictionary<string, IValue> plainValue)
-        {
-            avatarAddress = plainValue["avatarAddress"].ToAddress();
-        }
-        
-        
-        [Serializable]
-        public class DailyRewardResult : AttachmentActionResult
-        {
-            public Dictionary<Material, int> materials;
-            public Guid id;
-
-            protected override string TypeId => "dailyReward.dailyRewardResult";
-
-            public DailyRewardResult()
-            {
-            }
-
-            public DailyRewardResult(Bencodex.Types.Dictionary serialized) : base(serialized)
-            {
-                materials = serialized["materials"].ToDictionary_Material_int();
-                id = serialized["id"].ToGuid();
-            }
-
-            public override IValue Serialize() =>
-#pragma warning disable LAA1002
-                new Bencodex.Types.Dictionary(new Dictionary<IKey, IValue>
-                {
-                    [(Text) "materials"] = materials.Serialize(),
-                    [(Text) "id"] = id.Serialize(),
-                }.Union((Bencodex.Types.Dictionary) base.Serialize()));
-#pragma warning restore LAA1002
-        }
-    }
-}
